pyproject.toml
[project]
name = "manifold"
version = "0.1.0"
description = "Proxy mesh gateway — chain multiple LLM proxy services into a single transparent pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "typer>=0.12",
    "uvicorn>=0.30",
    "starlette>=0.38",
    "mcp>=1.0",
]

[project.scripts]
manifold = "manifold.cli:app"
manifold-mcp = "manifold.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/manifold"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.1.0",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src/manifold"]
omit = ["src/manifold/mcp_server.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 55
show_missing = true
